chore: move toWebRequest function to utils file
